+++ nvcodec/samples/src/main/java/org/bytedeco/nvcodec/samples/util/Dimension.java
@@ -1,0 +1,31 @@
+package org.bytedeco.nvcodec.samples.util;
+
+public class Dimension {
+    private int width;
+    private int height;
+
+    public Dimension() {
+        this(0, 0);
+    }
+
+    public Dimension(int width, int height) {
+        this.width = width;
+        this.height = height;
+    }
+
+    public int getWidth() {
+        return width;
+    }
+
+    public void setWidth(int width) {
+        this.width = width;
+    }
+
+    public int getHeight() {
+        return height;
+    }
+
+    public void setHeight(int height) {
+        this.height = height;
+    }
+}